feat: enhance AppsScreen with category selection and bottom sheet integration
