+++ src/Screens/Flows/App/AppsScreen/AppsScreen.tsx
@@ -1,14 +1,14 @@
 import { useNavigation } from "@react-navigation/native"
-import React, { useCallback } from "react"
+import React, { useCallback, useState } from "react"
 import { StyleSheet } from "react-native"
 import { BaseIcon, BaseTouchable, BaseSpacer, BaseView, HeaderStyleV2, HeaderTitle, Layout } from "~Components"
 import { COLORS, ColorThemeType } from "~Constants"
-import { useBottomSheetModal, useThemedStyles } from "~Hooks"
+import { useBottomSheetModal, useThemedStyles, useVeBetterDaoDapps } from "~Hooks"
 import { useIsNormalUser } from "~Hooks/useIsNormalUser"
 import { useI18nContext } from "~i18n"
 import { Routes } from "~Navigation"
 import { EcosystemSection } from "./Components/Ecosystem"
-import { VeBetterSection } from "./Components/VeBetter"
+import { AppsBottomSheet, VeBetterSection } from "./Components/VeBetter"
 import { ForYouCarousel } from "./Components/ForYouCarousel/ForYouCarousel"
 import { NewUserForYouCarousel } from "./Components/ForYouCarousel/NewUserForYouCarousel"
 import { useAppSelector } from "~Storage/Redux/Hooks"
@@ -18,13 +18,22 @@ import { Favourites } from "../DiscoverScreen/Components/Favourites"
 import { FavoritesBottomSheet } from "./Components/FavoritesBottomSheet"
 import { FavoritesSuggestionBanner } from "./Components/FavoritesSuggestionBanner"
 import { VeBetterDAOCarousel } from "../DiscoverScreen/Components/VeBetterDAOCarousel"
+import { X2ECategoryType } from "~Model"
 
 export const AppsScreen = () => {
     const { LL } = useI18nContext()
     const { styles, theme } = useThemedStyles(baseStyles)
     const nav = useNavigation()
-    const { ref: favoritesRef, onOpen: onOpenFavorites, onClose: onCloseFavorites } = useBottomSheetModal()
+    const [selectedCategoryId, setSelectedCategoryId] = useState<X2ECategoryType | undefined>()
 
+    const { ref: favoritesRef, onOpen: onOpenFavorites, onClose: onCloseFavorites } = useBottomSheetModal()
+    const {
+        ref: appsBottomSheetRef,
+        onOpen: onOpenAppsBottomSheet,
+        onClose: onCloseAppsBottomSheet,
+    } = useBottomSheetModal()
+
+    const { data: allApps, isLoading } = useVeBetterDaoDapps()
     const bookmarkedDApps = useAppSelector(selectBookmarkedDapps)
     const { onDAppPress } = useDAppActions()
 
@@ -35,6 +44,24 @@ export const AppsScreen = () => {
     const goToSearch = useCallback(() => {
         nav.navigate(Routes.APPS_SEARCH)
     }, [nav])
+
+    const handleCategoryPress = useCallback(
+        (categoryId: X2ECategoryType) => {
+            setSelectedCategoryId(categoryId)
+            onOpenAppsBottomSheet()
+        },
+        [onOpenAppsBottomSheet],
+    )
+
+    const handleCloseAppsBottomSheet = useCallback(() => {
+        onCloseAppsBottomSheet()
+        setSelectedCategoryId(undefined)
+    }, [onCloseAppsBottomSheet])
+
+    const handleFavoritesSuggestionPress = useCallback(() => {
+        setSelectedCategoryId(X2ECategoryType.NUTRITION)
+        onOpenAppsBottomSheet()
+    }, [onOpenAppsBottomSheet])
 
     return (
         <Layout
@@ -78,12 +105,12 @@ export const AppsScreen = () => {
 
                     <BaseSpacer height={48} />
 
-                    <VeBetterSection />
+                    <VeBetterSection onPressCategory={handleCategoryPress} />
                     <BaseSpacer height={48} />
 
                     {!showFavorites && (
                         <>
-                            <FavoritesSuggestionBanner onPress={() => {}} />
+                            <FavoritesSuggestionBanner onPress={handleFavoritesSuggestionPress} />
                             <BaseSpacer height={48} />
                         </>
                     )}
@@ -97,6 +124,13 @@ export const AppsScreen = () => {
 
                     <EcosystemSection />
                     <FavoritesBottomSheet ref={favoritesRef} onClose={onCloseFavorites} />
+                    <AppsBottomSheet
+                        ref={appsBottomSheetRef}
+                        onDismiss={handleCloseAppsBottomSheet}
+                        allApps={allApps}
+                        isLoading={isLoading}
+                        initialCategoryId={selectedCategoryId}
+                    />
                 </>
             }
         />
